ice: trade dyn_only only as suggestion
It can be needed to have some MSI-X allocated as static and rest as
dynamic. For example on PF VSI. We want to always have minimum one MSI-X
on it, because of that it is allocated as a static one, rest can be
dynamic if it is supported.

Change the ice_get_irq_res() to allow using static entries if they are
free even if caller wants dynamic one.

Fix limit values. Min and max in limit means the values that are valid,
so decrease max and num_static by one.

Set vsi::irq_dyn_alloc if dynamic allocation is supported.

Signed-off-by: Michal Swiatkowski <michal.swiatkowski@linux.intel.com>