ice: introduce uplink port representor
Uplink port representor is used to steer slow-path traffic. Tx from port
representors netdev will be done via uplink port representor netdev, Rx
will be done from uplink port representor netdev napi.

Add another type of PR using the same scheme as for VFs and subfunction
netdevs. Uplink port representor is a special one, because there can be
only one and it is always created when switching to switchdev mode.
Because of that store it as separate pointer instead of tracking it in
existing xarray.

Signed-off-by: Michal Swiatkowski <michal.swiatkowski@linux.intel.com>